requirements.txt
PyMuPDF
openai
google-generativeai